+++ Framework/Built_In_Automation/Sequential_Actions/sequential_actions.py
@@ -1972,13 +1972,18 @@ def Action_Handler(_data_set, action_row):
         for location in subprocess.getoutput("where python").split("\n"):
 	        if "Microsoft" not in location:
 		        python_folder.append(location)
+        try:
+            python_location = "" if len(python_folder) == 0 else "by going to {}".format(python_folder[0].split("Python")[0] + "Python")
+        except:
+            python_location = ""
         if not 3.5 <= float(sys.version.split(" ")[0][0:3]) <= 3.8:
             error_msg = "You have the wrong Python version or bit"\
                 +"\nFollow this procedure"\
                     +"\n1.Go to settings, then go to Apps and in search box type python and uninstall all python related things"\
-                        +"\n2.Delete your Python folder  by going to {}".format(python_folder[0].split("Python")[0] + "Python")\
+                        +"\n2.Delete your Python folder"\
+                            + python_location \
                             +"\n3.Go to this link and download python https://www.python.org/ftp/python/3.8.10/python-3.8.10-amd64.exe"\
-                                +"\n4.During installation give check 'Add Python to Path' and uncheck 'for all user'. This is very important."\
+                                +"\n4.During installation, give uncheck 'for all user' and check 'Add Python to Path'. This is very important."\
                                     +"\n5.Relaunch zeuz node_cli.py"
             CommonUtil.ExecLog(sModuleInfo, error_msg, 3,)
             return "zeuz_failed"