+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
 # Version 14.x.x
-## Version 14.4.x
+## Version 14.5.x
+
+### [14.5.2] [October 19, 2021]
+- **[Add]** Dictionary support is added in compare variable action
+
+### [14.5.1] [October 18, 2021]
+- **[Add]** key-value structure is added in excel_read action
 
 ### [14.5.0] [October 13, 2021]
 
@@ -16,6 +22,7 @@
 - **[Add]** Read and save mail action is added
 - **[Add]** Delete mail action is added
 - **[Add]** Search and save text using regex action is added
+## Version 14.4.x
 ### [14.4.0] [August 24, 2021]
 - **[Fix]** Fixed variable parsing limitations
 - **[Fix]** Fixed exact match bug and improved subset in compare_data action